+++ Runtime/RateManager.cs
@@ -3,6 +3,12 @@ using UnityEngine;
 using UniRate.Internals;
 using UniRate.Debug;
 
+#if UNITY_2021_1_OR_NEWER
+using UnityScreen = UnityEngine.Device.Screen;
+#else
+using UnityScreen = UnityEngine.Screen;
+#endif
+
 namespace UniRate {
 
     [DisallowMultipleComponent]
@@ -157,11 +163,11 @@ namespace UniRate {
                 this._guiContentText.text = $"Update/s: {this.UpdateRate.Current.ToString("000")} / {this.UpdateRate.Target.ToString("000")}\nFixedUpdate/s: {this.FixedUpdateRate.Current.ToString("000")} / {this.FixedUpdateRate.Target.ToString("000")}\nRenderInterval: {this.RenderInterval.Current.ToString()} / {this.RenderInterval.Target.ToString()}\nRender/s: {this.RenderInterval.CurrentRenderRate.ToString("000")}";
             }
 
-            var safeArea = Screen.safeArea;
+            var safeArea = UnityScreen.safeArea;
             RateDebug.ScreenDataStyle.CalcMinMaxWidth(this._guiContentText, out float minWidth, out float maxWidth);
             var labelWidth = Mathf.Min(safeArea.width - 10, maxWidth);
             float labelSpaceH = Mathf.Min(5f, ((safeArea.width - labelWidth) / 2f));
-            float labelSpaceV = Math.Max(Screen.height - safeArea.height, 5f);
+            float labelSpaceV = Math.Max(UnityScreen.height - safeArea.height, 5f);
             var rectLine = new Rect(safeArea.x + labelSpaceH, labelSpaceV, labelWidth, RateDebug.ScreenDataStyle.CalcHeight(this._guiContentText, labelWidth));
 
             if (GUI.Button(rectLine, this._guiContentText, RateDebug.ScreenDataStyle)) {